Move mocks to different place
